djangotutorial/requirements.txt
Django==4.2.20  # Последняя стабильная версия Django 4.2.x (на момент ответа)
psycopg2-binary==2.9.9
djangorestframework==3.14.0